pyproject.toml
[tool.poetry]
name = "mospolytech-tg-bot"
version = "0.1.0"
description = ""
authors = ["Artem Zlobin <arzar0046@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.13.1"
sqlalchemy = "^2.0.35"
pydantic-settings = "^2.5.2"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
celery = "^5.4.0"
redis = "^5.1.1"

[tool.black]
line-length=79
target-version=['py39']
include='app'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
